02/22

+++ javaProject/src/variables/GetAverageValue.java
@@ -6,12 +6,12 @@ public class GetAverageValue {
 		int a = 17;
 		int b = 18;
 		
-		double c = (double) (a + b)/2;
+		double c =  (a + b)/2.0;
 		
 		System.out.println("두수의 평균은 : " +c + "입니다.");
+	 
 		
-		
-
+ 
 	}
 
 }
